feat(kv-router): extract scheduling logic from lib/llm into lib/kv-router
Move config, scheduling types, worker selector, and queue into
dynamo-kv-router so the scheduling layer is benchmarkable without
pulling in dynamo-runtime. lib/llm re-exports everything via thin
type aliases—zero runtime overhead, all monomorphized.
