+++ lib/kv-router/Cargo.toml
@@ -24,6 +24,7 @@ dynamo-tokens = { workspace = true }
 anyhow = { workspace = true }
 async-trait = { workspace = true }
 dashmap = { workspace = true }
+derive_builder = { workspace = true }
 derive-getters = { workspace = true }
 prometheus = { workspace = true }
 rand = { workspace = true }
@@ -34,6 +35,7 @@ tokio = { workspace = true }
 tokio-util = { workspace = true }
 tracing = { workspace = true }
 uuid = { workspace = true }
+validator = { workspace = true }
 xxhash-rust = { workspace = true }
 
 # dependencies
