+++ lib/llm/src/kv_router/scheduler.rs
@@ -1,10 +1,15 @@
 // SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+pub use dynamo_kv_router::scheduling::{
+    KvSchedulerError, PotentialLoad, SchedulingRequest, SchedulingResponse,
+};
+pub use dynamo_kv_router::selector::DefaultWorkerSelector;
+
 use super::KvRouterConfig;
 use super::RouterConfigOverride;
 use super::WorkerSelector;
-use super::protocols::{DpRank, OverlapScores, WorkerId, WorkerSelectionResult, WorkerWithDpRank};
+use super::protocols::{OverlapScores, WorkerId};
 use super::queue::SchedulerQueue;
 use super::sequence::{
     ActiveSequencesMulti, SequenceError, SequenceRequest, create_multi_worker_sequences,
@@ -14,8 +19,6 @@ use crate::local_model::runtime_config::ModelRuntimeConfig;
 use anyhow::Result;
 use dynamo_runtime::component::Component;
 use dynamo_runtime::traits::DistributedRuntimeProvider;
-use rand::Rng;
-use serde::{Deserialize, Serialize};
 use std::collections::{HashMap, HashSet};
 use std::sync::Arc;
 use std::time::Duration;
@@ -23,64 +26,6 @@ use std::time::Duration;
 use std::time::Instant;
 
 use dynamo_tokens::SequenceHash;
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PotentialLoad {
-    pub worker_id: WorkerId,
-    pub dp_rank: DpRank,
-    pub potential_prefill_tokens: usize,
-    pub potential_decode_blocks: usize,
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum KvSchedulerError {
-    #[error("no endpoints available to route work")]
-    NoEndpoints,
-
-    #[error("endpoint subscriber shutdown")]
-    SubscriberShutdown,
-
-    #[error("failed to initialize event publisher: {0}")]
-    InitFailed(String),
-}
-
-#[derive(Debug)]
-pub struct SchedulingResponse {
-    pub best_worker: WorkerWithDpRank,
-    pub overlap_blocks: u32,
-}
-
-pub struct SchedulingRequest {
-    pub maybe_request_id: Option<String>,
-    pub token_seq: Option<Vec<SequenceHash>>,
-    pub isl_tokens: usize,
-    pub overlaps: OverlapScores,
-    pub decode_blocks: HashMap<WorkerWithDpRank, usize>,
-    pub prefill_tokens: HashMap<WorkerWithDpRank, usize>,
-    // Router config overrides for this specific request
-    pub router_config_override: Option<RouterConfigOverride>,
-    // Whether to update scheduler states (false for query_instance_id requests)
-    pub update_states: bool,
-    // LORA adapter name extracted from request.model field
-    pub lora_name: Option<String>,
-    /// Priority jump in seconds; decreases effective arrival time in the queue.
-    pub priority_jump: f64,
-    /// Optional set of allowed worker IDs to restrict routing decisions (EPP).
-    pub allowed_worker_ids: Option<HashSet<WorkerId>>,
-    resp_tx: Option<tokio::sync::oneshot::Sender<Result<SchedulingResponse, KvSchedulerError>>>,
-}
-
-impl SchedulingRequest {
-    pub fn respond(&mut self, result: Result<SchedulingResponse, KvSchedulerError>) {
-        let Some(tx) = self.resp_tx.take() else {
-            tracing::error!("respond called multiple times on same request");
-            return;
-        };
-        if tx.send(result).is_err() {
-            tracing::error!("failed to send response to requestor");
-        }
-    }
-}
 
 pub struct KvScheduler {
     request_tx: tokio::sync::mpsc::Sender<SchedulingRequest>,
@@ -93,7 +38,7 @@ impl KvScheduler {
         component: Component,
         block_size: u32,
         workers_with_configs: RuntimeConfigWatch,
-        selector: Option<Box<dyn WorkerSelector + Send + Sync>>,
+        selector: Option<Box<WorkerSelector>>,
         kv_router_config: &KvRouterConfig,
         worker_type: &'static str,
     ) -> Result<Self, KvSchedulerError> {
@@ -300,7 +245,7 @@ impl KvScheduler {
                 .potential_blocks_and_tokens(token_seq.as_deref(), isl_tokens, overlaps);
 
         // Get all unique WorkerWithDpRank from both hashmaps
-        let mut workers: HashSet<WorkerWithDpRank> = HashSet::new();
+        let mut workers: HashSet<dynamo_kv_router::protocols::WorkerWithDpRank> = HashSet::new();
         workers.extend(decode_blocks.keys().copied());
         workers.extend(prefill_tokens.keys().copied());
 
@@ -324,316 +269,5 @@ impl KvScheduler {
     /// Get active request counts grouped by LORA name
     pub fn get_active_lora_counts(&self) -> HashMap<String, usize> {
         self.slots.get_active_lora_counts()
-    }
-}
-
-// Helper function for softmax sampling
-// Returns a vec of workers: multiple if tied, single if sampled
-fn softmax_sample(
-    logits: &HashMap<WorkerWithDpRank, f64>,
-    temperature: f64,
-) -> Vec<WorkerWithDpRank> {
-    if logits.is_empty() {
-        panic!("Empty logits for softmax sampling");
-    }
-
-    // Guard: if temperature is 0, return all keys with the smallest logit value (ties)
-    if temperature == 0.0 {
-        // Find the minimum logit value
-        let min_logit = logits.values().fold(f64::INFINITY, |a, &b| a.min(b));
-
-        // Collect all keys with the minimum logit value (to handle ties)
-        let min_keys: Vec<_> = logits
-            .iter()
-            .filter(|&(_, &v)| v == min_logit)
-            .map(|(k, _)| *k)
-            .collect();
-
-        return min_keys;
-    }
-
-    let keys: Vec<_> = logits.keys().copied().collect();
-    let values: Vec<_> = logits.values().copied().collect();
-
-    // Find min and max for normalization
-    let min_val = values.iter().fold(f64::INFINITY, |a, &b| a.min(b));
-    let max_val = values.iter().fold(f64::NEG_INFINITY, |a, &b| a.max(b));
-
-    let probabilities = if min_val == max_val {
-        // All values are the same, uniform probability
-        vec![1.0 / keys.len() as f64; keys.len()]
-    } else {
-        // Fused normalize → negate → scale → exp, then normalize probabilities
-        let range = max_val - min_val;
-        let scaled: Vec<f64> = values.iter().map(|&v| -(v / range) / temperature).collect();
-        let max_scaled = scaled.iter().fold(f64::NEG_INFINITY, |a, &b| a.max(b));
-        let mut probs: Vec<f64> = scaled.iter().map(|&v| (v - max_scaled).exp()).collect();
-        let sum: f64 = probs.iter().sum();
-        probs.iter_mut().for_each(|p| *p /= sum);
-        probs
-    };
-
-    // Sample from the probability distribution
-    let mut rng = rand::rng();
-    let sample: f64 = rng.random();
-
-    let mut cumsum = 0.0;
-    for (i, &prob) in probabilities.iter().enumerate() {
-        cumsum += prob;
-        if sample <= cumsum {
-            return vec![keys[i]];
-        }
-    }
-
-    // Fallback to last key (shouldn't normally reach here)
-    vec![keys[keys.len() - 1]]
-}
-
-// Default implementation matching the Python _cost_function
-#[derive(Debug, Clone, Default)]
-pub struct DefaultWorkerSelector {
-    pub kv_router_config: KvRouterConfig,
-}
-
-impl DefaultWorkerSelector {
-    pub fn new(kv_router_config: Option<KvRouterConfig>) -> Self {
-        Self {
-            kv_router_config: kv_router_config.unwrap_or_default(),
-        }
-    }
-}
-
-impl WorkerSelector for DefaultWorkerSelector {
-    fn select_worker(
-        &self,
-        workers: &HashMap<WorkerId, ModelRuntimeConfig>,
-        request: &SchedulingRequest,
-        block_size: u32,
-    ) -> Result<WorkerSelectionResult, KvSchedulerError> {
-        assert!(request.isl_tokens > 0);
-
-        let allowed_ids = request.allowed_worker_ids.as_ref();
-
-        if allowed_ids.map_or(workers.is_empty(), |ids| {
-            !workers.keys().any(|wid| ids.contains(wid))
-        }) {
-            return Err(KvSchedulerError::NoEndpoints);
-        }
-
-        let isl = request.isl_tokens;
-        let request_blocks = isl.div_ceil(block_size as usize);
-        let overlaps = &request.overlaps.scores;
-
-        let decode_blocks = &request.decode_blocks;
-        let prefill_tokens = &request.prefill_tokens;
-
-        let mut worker_logits = HashMap::new();
-
-        // Use override if provided, otherwise use default config
-        let overlap_weight = request
-            .router_config_override
-            .as_ref()
-            .and_then(|cfg| cfg.overlap_score_weight)
-            .unwrap_or(self.kv_router_config.overlap_score_weight);
-
-        for (worker_id, config) in workers
-            .iter()
-            .filter(|(wid, _)| allowed_ids.is_none_or(|ids| ids.contains(wid)))
-        {
-            let data_parallel_size = config.data_parallel_size;
-
-            for dp_rank in 0..data_parallel_size {
-                let worker = WorkerWithDpRank::new(*worker_id, dp_rank);
-
-                // Get overlap for this worker (defaults to 0 if not in overlaps)
-                let overlap = *overlaps.get(&worker).unwrap_or(&0);
-
-                // this is the number of prefill tokens the worker would have if the request were scheduled there
-                let prefill_token = *prefill_tokens.get(&worker).unwrap_or(&isl);
-                let potential_prefill_block = (prefill_token as f64) / (block_size as f64);
-
-                // this is the number of decode blocks the worker would have if the request were scheduled there
-                let decode_block = *decode_blocks
-                    .get(&worker)
-                    .unwrap_or(&(potential_prefill_block.floor() as usize))
-                    as f64;
-
-                // Calculate logit (lower is better)
-                let logit = overlap_weight * potential_prefill_block + decode_block;
-
-                worker_logits.insert(worker, logit);
-
-                tracing::info!(
-                    "Formula for worker_id={} dp_rank={:?} with {overlap} cached blocks: {logit:.3} \
-                     = {overlap_weight:.1} * prefill_blocks + decode_blocks \
-                     = {overlap_weight:.1} * {potential_prefill_block:.3} + {decode_block:.3}",
-                    worker.worker_id,
-                    worker.dp_rank
-                );
-            }
-        }
-
-        // Use softmax sampling to select worker(s)
-        // Use override if provided, otherwise use default config
-        let temperature = request
-            .router_config_override
-            .as_ref()
-            .and_then(|cfg| cfg.router_temperature)
-            .unwrap_or(self.kv_router_config.router_temperature);
-        let candidates = softmax_sample(&worker_logits, temperature);
-
-        // If multiple candidates (tied), use tree size as tie-breaker
-        // If tree sizes are also equal, use random selection to avoid bias
-        let best_worker = if candidates.len() > 1 {
-            tracing::info!("Multiple workers tied with same logit, using tree size as tie-breaker");
-            let tree_sizes: Vec<(usize, &WorkerWithDpRank)> = candidates
-                .iter()
-                .map(|w| (request.overlaps.tree_sizes.get(w).copied().unwrap_or(0), w))
-                .collect();
-
-            if tree_sizes.iter().all(|(s, _)| *s == tree_sizes[0].0) {
-                let idx = rand::rng().random_range(0..candidates.len());
-                candidates[idx]
-            } else {
-                *tree_sizes.iter().min_by_key(|(s, _)| *s).unwrap().1
-            }
-        } else {
-            candidates[0]
-        };
-
-        let best_logit = worker_logits[&best_worker];
-
-        let best_overlap = *overlaps.get(&best_worker).unwrap_or(&0);
-
-        // this is a runtime config set on a per worker basis, not per dp-rank
-        let total_blocks_info = workers
-            .get(&best_worker.worker_id)
-            .and_then(|cfg| cfg.total_kv_blocks)
-            .map(|blocks| format!(", total blocks: {}", blocks))
-            .unwrap_or_default();
-
-        let tree_size = request
-            .overlaps
-            .tree_sizes
-            .get(&best_worker)
-            .copied()
-            .unwrap_or(0);
-
-        tracing::info!(
-            "Selected worker: worker_id={} dp_rank={:?}, logit: {:.3}, cached blocks: {}, tree size: {}{}",
-            best_worker.worker_id,
-            best_worker.dp_rank,
-            best_logit,
-            best_overlap,
-            tree_size,
-            total_blocks_info
-        );
-
-        Ok(WorkerSelectionResult {
-            worker: best_worker,
-            required_blocks: request_blocks as u64,
-            overlap_blocks: overlaps.get(&best_worker).copied().unwrap_or(0),
-        })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_softmax_sample_single_key() {
-        // Test that with a single key, softmax_sample always returns that key
-        let mut logits = HashMap::new();
-        let worker = WorkerWithDpRank::from_worker_id(42);
-        logits.insert(worker, 0.5); // The value doesn't matter
-
-        // Test with different temperatures
-        for temperature in &[0.1, 1.0, 10.0] {
-            let result = softmax_sample(&logits, *temperature);
-            assert_eq!(result.len(), 1, "Should return exactly one worker");
-            assert_eq!(result[0], worker, "Should return the only available worker");
-        }
-
-        // Test with different logit values
-        logits.clear();
-        logits.insert(worker, -100.0); // Very negative value
-        let result = softmax_sample(&logits, 1.0);
-        assert_eq!(result.len(), 1);
-        assert_eq!(result[0], worker);
-
-        logits.clear();
-        logits.insert(worker, 100.0); // Very positive value
-        let result = softmax_sample(&logits, 1.0);
-        assert_eq!(result.len(), 1);
-        assert_eq!(result[0], worker);
-
-        logits.clear();
-        logits.insert(worker, 0.0); // Zero value
-        let result = softmax_sample(&logits, 1.0);
-        assert_eq!(result.len(), 1);
-        assert_eq!(result[0], worker);
-    }
-
-    #[test]
-    fn test_softmax_sample_zero_temperature() {
-        // Test that with temperature 0, softmax_sample returns all keys with smallest logit
-        let mut logits = HashMap::new();
-        let worker1 = WorkerWithDpRank::from_worker_id(1);
-        let worker2 = WorkerWithDpRank::from_worker_id(2);
-        let worker3 = WorkerWithDpRank::from_worker_id(3);
-        let worker4 = WorkerWithDpRank::from_worker_id(4);
-        logits.insert(worker1, 5.0);
-        logits.insert(worker2, 3.0); // This has the smallest logit
-        logits.insert(worker3, 7.0);
-        logits.insert(worker4, 3.5);
-
-        // With temperature 0, should always return only worker2 (smallest logit)
-        let result = softmax_sample(&logits, 0.0);
-        assert_eq!(
-            result.len(),
-            1,
-            "Should return one worker when there's no tie"
-        );
-        assert_eq!(
-            result[0], worker2,
-            "Should return worker with smallest logit when temperature is 0"
-        );
-
-        // Test with tied minimum logits
-        logits.clear();
-        let worker5 = WorkerWithDpRank::from_worker_id(5);
-        let worker6 = WorkerWithDpRank::from_worker_id(6);
-        logits.insert(worker1, 5.0);
-        logits.insert(worker2, 3.0); // Tied for smallest
-        logits.insert(worker5, 3.0); // Tied for smallest
-        logits.insert(worker6, 7.0);
-
-        let result = softmax_sample(&logits, 0.0);
-        assert_eq!(
-            result.len(),
-            2,
-            "Should return all workers with smallest logit when tied"
-        );
-        assert!(
-            result.contains(&worker2) && result.contains(&worker5),
-            "Should contain both tied workers"
-        );
-
-        // Test with negative values
-        logits.clear();
-        let worker10 = WorkerWithDpRank::from_worker_id(10);
-        let worker20 = WorkerWithDpRank::from_worker_id(20);
-        let worker30 = WorkerWithDpRank::from_worker_id(30);
-        logits.insert(worker10, -1.0);
-        logits.insert(worker20, -5.0); // This has the smallest logit
-        logits.insert(worker30, 0.0);
-
-        let result = softmax_sample(&logits, 0.0);
-        assert_eq!(result.len(), 1);
-        assert_eq!(
-            result[0], worker20,
-            "Should handle negative logits correctly"
-        );
     }
 }
